requirements.txt
python-dotenv
pydantic
openai

GitPython